refactor: 요청, 응답에 대한 역할을 controller 계층에 위임
