+++ src/main/java/roomescape/login/AuthController.java
@@ -3,6 +3,7 @@ package roomescape.login;
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.util.Arrays;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -32,7 +33,18 @@ public class AuthController {
 
     @GetMapping("/login/check")
     public ResponseEntity<MemberResponse> checkLogin(HttpServletRequest request) {
-        MemberResponse memberResponse = authService.checkLogin(request);
+        Cookie[] cookies = request.getCookies();
+        String token = extractTokenFromCookie(cookies);
+        MemberResponse memberResponse = authService.checkLogin(token);
+
         return ResponseEntity.ok(memberResponse);
+    }
+
+    private String extractTokenFromCookie(Cookie[] cookies) {
+        return Arrays.stream(cookies)
+                .filter(cookie -> cookie.getName().equals("token"))
+                .map(Cookie::getValue)
+                .findFirst()
+                .orElse("");
     }
 }
