+++ src/main/java/roomescape/login/AuthService.java
@@ -25,20 +25,10 @@ public class AuthService {
         return new AuthResponse(accessToken);
     }
 
-    public MemberResponse checkLogin(HttpServletRequest request) {
-        Cookie[] cookies = request.getCookies();
-        String token = extractTokenFromCookie(cookies);
+    public MemberResponse checkLogin(String token) {
         Long memberId = jwtTokenProvider.getMemberId(token);
         Member member = memberDao.findById(memberId);
 
         return new MemberResponse(member.getId(), member.getName(), member.getEmail());
-    }
-
-    private String extractTokenFromCookie(Cookie[] cookies) {
-        return Arrays.stream(cookies)
-                .filter(cookie -> cookie.getName().equals("token"))
-                .map(Cookie::getValue)
-                .findFirst()
-                .orElse("");
     }
 }
